Remove duplicate title attributes
I'm not sure why this was done. As far as I'm aware of this
_creates_ an accessibility issue. The same text is read twice by
screenreader software, for example. It's also a problem for users
that can see because they see a tooltip and have to read it just
to learn that there is nothing new to learn in the tooltip.

Change-Id: I4bbf97595b8be1b278ad20305935d14602457213