Remove dead code from fullscreen map code
This is dead code since 2016, see Id98f23a, tracked via T144880. What
this code was meant to do is to update the URL in the browser's
address bar. I would argue this was not a useful feature to begin
with. When I, as a user, look at a map that belongs to an article and
I want to share a URL, what's more likely? Thart I want to share the
map as shown in the article, or something else?

Bug: T144880
Change-Id: Ib25d01adcfe55ce703b60c1efd407369315bcd1e