PHP 8.1 compatibility: Don't pass null to string functions
trim() doesn't silently accept null any more since PHP 8.1.
https://www.php.net/manual/en/migration81.deprecated.php#migration81.deprecated.core.null-not-nullable-internal

Same issue as in T289926 and T313663, but there is no umbrella ticket
to assign this one to.
https://phabricator.wikimedia.org/tag/php_8.1_support/

I use the opportunity to re-arrange the code a little bit. We pass
the untrimmed string to the JSON decoder now. I tested this and it
works.

Change-Id: I8171170104216a3a6d93321757a2d0b668ae3537