Replace data-mw='interface' with usage of data-mw-kartographer
Native support of Kartographer tags by Parsoid induces a conflict over
the use of data-mw, which was so far used in the pattern
"data-mw='interface'" to avoid HTML map injection directly in the
wikitext. The goal of this patch is to eventually stop the usage of this
data-mw='interface' pattern and to replace it with checking for the
presence of a data-mw-kartographer attribute instead.

This patch temporarily maintains the compatibility to the older version
and checks for both data-mw='interface' and the presence of
data-mw-kartographer to avoid HTML injection in the JavaScript so that
cached version of HTML does not break. A follow-up patch, once the caches
are clear, will remove the data-mw='interface' check to only keep the
check for data-mw-kartographer.

The data-mw-kartographer attribute aims at staying empty ("marker
attribute) for the "legacy" version of the Kartographer tag rendering,
and at containing Kartographer-related information in the future
"Parsoid" version of the rendering.

Bug: T331652
Depends-On: I5d52cddb3d672de0130f7a39b25f3fbbf09ac250
Change-Id: If163bf0b29f5669336138dee746403e8107559bb